ai_service/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
sentence-transformers
transformers
torch
thefuzz
scikit-learn
sqlalchemy
psycopg2-binary
alembic
python-dotenv

PyMuPDF
google-genai
spacy
